+++ tarantool/tlua/src/lib.rs
@@ -167,6 +167,8 @@ mod lua_tables;
 mod macros;
 mod object;
 mod rust_tables;
+#[cfg(feature = "extra_impls")]
+mod smol_str;
 #[cfg(feature = "internal_test")]
 pub mod test;
 mod tuples;